# FreeTimeApp/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "tiempo-libre",
    "version": "1.0"
  },
  "paths": {
...
    "/api/vacaciones/disponibilidad": {
      "get": {
        "tags": [
          "ReservaVacaciones"
        ],
        "parameters": [
          {
            "name": "anio",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "grupoId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vacaciones/validar-disponibilidad/{empleadoId}": {
      "get": {
        "tags": [
          "ReservaVacaciones"
        ],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/vacaciones/reservar-anual": {
      "post": {
        "tags": [
          "ReservaVacaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaAnualRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaAnualRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaAnualRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReservaAnualRequest": {
        "required": [
          "anioVacaciones",
          "empleadoId",
          "fechasSeleccionadas"
        ],
        "type": "object",
        "properties": {
          "empleadoId": {
            "type": "integer",
            "format": "int32"
          },
          "anioVacaciones": {
            "type": "integer",
            "format": "int32"
          },
          "fechasSeleccionadas": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          },
          "observaciones": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}